approaching final regiment charge dev
Prioritization of the regiments made, with for each one a path.
We just need to display the path on the selected regiment.

+++ Unity/Assets/Scripts/COMBAT SCRIPTS/CombatVariables.cs
@@ -5,8 +5,10 @@ using UnityEngine;
 public class CombatVariables : MonoBehaviour
 {
     //variables for the game design
-    public float movesPerRound = 1;
-    public float moveCapacity = 1;
+    //Moves per round is the standard move per round autorisation for this unit (not modified except by winning experience tier etc)
+    public int movesPerRound = 1;
+    //MoveCapacity is the actual move capacity for the unit, so it can be less or more depending on actual buffers
+    public int moveCapacity = 1;
     public int range = 1;
     public int rangeHeight = 2;
     public int people = 3;
